refactor(#386): extract shared B3.Exchange.TestSupport for ResolveRepoFile/TryDeleteDir
Per architectural review 2026-05-20 (refactor #7): the same temp-dir and
repo-file bootstrap was duplicated across 19 test files in 4 test
projects. When the bootstrap changes, only some sites get updated and
the suite drifts.

Extract into new tests/B3.Exchange.TestSupport library:
- TestPaths.ResolveRepoFile(string) — walk up from AppContext.BaseDirectory
- TempDirs.TryDelete(string) — best-effort recursive delete
- TempDirs.Create(string prefix) — fresh unique temp dir (available for
  future migration of scattered Guid-based temp-dir patterns)

Migrated 19 test files to consume the shared helpers and deleted their
private duplicates. No behavioral changes; build + test + format green.

Closes #386.
