+++ tests/B3.Exchange.Host.Tests/GracefulShutdownTests.cs
@@ -2,6 +2,7 @@ using System.Buffers.Binary;
 using System.Net.Sockets;
 using B3.EntryPoint.Wire;
 using B3.Exchange.Core;
+using B3.Exchange.TestSupport;
 
 namespace B3.Exchange.Host.Tests;
 
@@ -23,7 +24,7 @@ public class GracefulShutdownTests
 
     private static HostConfig BuildConfig()
     {
-        var instrumentsPath = ResolveRepoFile("config/instruments-eqt.json");
+        var instrumentsPath = TestPaths.ResolveRepoFile("config/instruments-eqt.json");
         return new HostConfig
         {
             Auth = new AuthConfig { RequireFixpHandshake = false },
@@ -43,17 +44,6 @@ public class GracefulShutdownTests
         };
     }
 
-    private static string ResolveRepoFile(string relPath)
-    {
-        var dir = AppContext.BaseDirectory;
-        for (int i = 0; i < 8 && dir != null; i++)
-        {
-            var candidate = Path.Combine(dir, relPath);
-            if (File.Exists(candidate)) return candidate;
-            dir = Path.GetDirectoryName(dir);
-        }
-        throw new FileNotFoundException($"could not locate {relPath} from {AppContext.BaseDirectory}");
-    }
 
     [Fact]
     public async Task StopAsync_BroadcastsTerminateFinished_BeforeTcpClose()
