+++ tests/B3.Exchange.Host.Tests/PostTradeBustE2ETests.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using B3.Exchange.Core;
 using B3.Exchange.Matching;
 using B3.Exchange.PostTrade;
+using B3.Exchange.TestSupport;
 
 namespace B3.Exchange.Host.Tests;
 
@@ -67,17 +68,6 @@ public sealed class PostTradeBustE2ETests : IDisposable
         public void Publish(byte channelNumber, ReadOnlySpan<byte> packet) { }
     }
 
-    private static string ResolveRepoFile(string relPath)
-    {
-        var dir = AppContext.BaseDirectory;
-        for (int i = 0; i < 8 && dir != null; i++)
-        {
-            var candidate = Path.Combine(dir, relPath);
-            if (File.Exists(candidate)) return candidate;
-            dir = Path.GetDirectoryName(dir);
-        }
-        throw new FileNotFoundException($"could not locate {relPath} from {AppContext.BaseDirectory}");
-    }
 
     private static PostTradeRecord MakeFill(uint id) => new(
         TradeId: id, TransactTimeNanos: TradeDateNanos + id * 1_000UL, SecurityId: Petr,
@@ -115,7 +105,7 @@ public sealed class PostTradeBustE2ETests : IDisposable
     [Fact]
     public async Task FullFlow_PreEodBust_FoldedOut_PostEodBust_PublishedToAmendmentsCsv()
     {
-        var instrumentsPath = ResolveRepoFile("config/instruments-eqt.json");
+        var instrumentsPath = TestPaths.ResolveRepoFile("config/instruments-eqt.json");
 
         // 1) Pre-seed three fills before host boot so the dedup index
         //    picks them up at startup and the validator can locate the
