+++ tests/B3.Exchange.Host.Tests/SoakSmokeTests.cs
@@ -4,6 +4,7 @@ using B3.Exchange.Gateway;
 using B3.Exchange.Core;
 using B3.Exchange.Matching;
 using B3.Exchange.SyntheticTrader;
+using B3.Exchange.TestSupport;
 
 namespace B3.Exchange.Host.Tests;
 
@@ -77,7 +78,7 @@ public class SoakSmokeTests
             return;
         }
 
-        var instrumentsPath = ResolveRepoFile("config/instruments-eqt.json");
+        var instrumentsPath = TestPaths.ResolveRepoFile("config/instruments-eqt.json");
         var sink = new MonotonicSink();
         var hostCfg = new HostConfig
         {
@@ -207,15 +208,4 @@ public class SoakSmokeTests
         }
     }
 
-    private static string ResolveRepoFile(string relPath)
-    {
-        var dir = AppContext.BaseDirectory;
-        for (int i = 0; i < 8 && dir != null; i++)
-        {
-            var candidate = Path.Combine(dir, relPath);
-            if (File.Exists(candidate)) return candidate;
-            dir = Path.GetDirectoryName(dir);
-        }
-        throw new FileNotFoundException($"could not locate {relPath} from {AppContext.BaseDirectory}");
-    }
 }
