+++ tests/B3.Exchange.Interop.Tests/EpcInteropTests.cs
@@ -5,6 +5,7 @@ using B3.EntryPoint.Client.Models;
 using B3.Exchange.Core;
 using B3.Exchange.Host;
 using Microsoft.Extensions.Logging;
+using B3.Exchange.TestSupport;
 
 namespace B3.Exchange.Interop.Tests;
 
@@ -27,17 +28,6 @@ public class EpcInteropTests : IAsyncLifetime
         public void Publish(byte channelNumber, ReadOnlySpan<byte> packet) => Interlocked.Increment(ref Count);
     }
 
-    private static string ResolveRepoFile(string relPath)
-    {
-        var dir = AppContext.BaseDirectory;
-        for (int i = 0; i < 8 && dir != null; i++)
-        {
-            var candidate = Path.Combine(dir, relPath);
-            if (File.Exists(candidate)) return candidate;
-            dir = Path.GetDirectoryName(dir);
-        }
-        throw new FileNotFoundException($"could not locate {relPath} from {AppContext.BaseDirectory}");
-    }
 
     private ExchangeHost _host = null!;
     private IPEndPoint _endpoint = null!;
@@ -85,7 +75,7 @@ public class EpcInteropTests : IAsyncLifetime
                     IncrementalGroup = "239.255.43.84",
                     IncrementalPort = port,
                     Ttl = 0,
-                    InstrumentsFile = ResolveRepoFile("config/instruments-eqt.json"),
+                    InstrumentsFile = TestPaths.ResolveRepoFile("config/instruments-eqt.json"),
                 },
             },
         };
