+++ tests/B3.Exchange.ScenarioReplay.Tests/MultiSessionReplayTests.cs
@@ -2,6 +2,7 @@ using B3.Exchange.Core;
 using B3.Exchange.Host;
 using B3.Exchange.SyntheticTrader;
 using B3.Exchange.SyntheticTrader.Fixp;
+using B3.Exchange.TestSupport;
 
 namespace B3.Exchange.ScenarioReplay.Tests;
 
@@ -22,17 +23,6 @@ public class MultiSessionReplayTests
         public void Publish(byte channelNumber, ReadOnlySpan<byte> packet) => Interlocked.Increment(ref Count);
     }
 
-    private static string ResolveRepoFile(string relPath)
-    {
-        var dir = AppContext.BaseDirectory;
-        for (int i = 0; i < 8 && dir != null; i++)
-        {
-            var candidate = Path.Combine(dir, relPath);
-            if (File.Exists(candidate)) return candidate;
-            dir = Path.GetDirectoryName(dir);
-        }
-        throw new FileNotFoundException($"could not locate {relPath} from {AppContext.BaseDirectory}");
-    }
 
     [Fact]
     public async Task TwoSessions_CrossingScript_LandsErTradesOnBothSides()
@@ -60,7 +50,7 @@ public class MultiSessionReplayTests
                     IncrementalGroup = "239.255.42.84",
                     IncrementalPort = 30192,
                     Ttl = 0,
-                    InstrumentsFile = ResolveRepoFile("config/instruments-eqt.json"),
+                    InstrumentsFile = TestPaths.ResolveRepoFile("config/instruments-eqt.json"),
                 },
             },
         };
@@ -206,7 +196,7 @@ public class MultiSessionReplayTests
                     IncrementalGroup = "239.255.42.84",
                     IncrementalPort = 30193,
                     Ttl = 0,
-                    InstrumentsFile = ResolveRepoFile("config/instruments-eqt.json"),
+                    InstrumentsFile = TestPaths.ResolveRepoFile("config/instruments-eqt.json"),
                 },
             },
         };
