+++ tests/B3.Exchange.SyntheticTrader.Tests/Fixp/FixpClientHandshakeTests.cs
@@ -1,6 +1,7 @@
 using B3.Exchange.Core;
 using B3.Exchange.Host;
 using B3.Exchange.SyntheticTrader.Fixp;
+using B3.Exchange.TestSupport;
 
 namespace B3.Exchange.SyntheticTrader.Tests.Fixp;
 
@@ -22,17 +23,6 @@ public class FixpClientHandshakeTests
         public void Publish(byte channelNumber, ReadOnlySpan<byte> packet) => Interlocked.Increment(ref Count);
     }
 
-    private static string ResolveRepoFile(string relPath)
-    {
-        var dir = AppContext.BaseDirectory;
-        for (int i = 0; i < 8 && dir != null; i++)
-        {
-            var candidate = Path.Combine(dir, relPath);
-            if (File.Exists(candidate)) return candidate;
-            dir = Path.GetDirectoryName(dir);
-        }
-        throw new FileNotFoundException($"could not locate {relPath} from {AppContext.BaseDirectory}");
-    }
 
     [Fact]
     public async Task FixpClient_HandshakeAndOrderRoundTrip()
@@ -58,7 +48,7 @@ public class FixpClientHandshakeTests
                     IncrementalGroup = "239.255.42.84",
                     IncrementalPort = 30190,
                     Ttl = 0,
-                    InstrumentsFile = ResolveRepoFile("config/instruments-eqt.json"),
+                    InstrumentsFile = TestPaths.ResolveRepoFile("config/instruments-eqt.json"),
                 },
             },
         };
@@ -126,7 +116,7 @@ public class FixpClientHandshakeTests
                     IncrementalGroup = "239.255.42.84",
                     IncrementalPort = 30191,
                     Ttl = 0,
-                    InstrumentsFile = ResolveRepoFile("config/instruments-eqt.json"),
+                    InstrumentsFile = TestPaths.ResolveRepoFile("config/instruments-eqt.json"),
                 },
             },
         };
